requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
openai
python-dotenv